+++ src/main/java/com/tanja/web_customer_tracker/DAO/customer/CustomerDAO.java
@@ -1,8 +1,8 @@
-package com.tanja.web_customer_tracker.DAO;
+package com.tanja.web_customer_tracker.DAO.customer;
 
 import java.util.List;
 
-import com.tanja.web_customer_tracker.model.Customer;
+import com.tanja.web_customer_tracker.model.customer.Customer;
 
 
 public interface CustomerDAO {
