+++ src/main/java/com/tanja/web_customer_tracker/DAO/project/ProjectDAO.java
@@ -1,8 +1,8 @@
-package com.tanja.web_customer_tracker.DAO;
+package com.tanja.web_customer_tracker.DAO.project;
 
 import java.util.List;
 
-import com.tanja.web_customer_tracker.model.Project;
+import com.tanja.web_customer_tracker.model.project.Project;
 
 public interface ProjectDAO {
