+++ src/main/java/com/tanja/web_customer_tracker/model/customer/GenderEnum.java
@@ -1,4 +1,4 @@
-package com.tanja.web_customer_tracker.model;
+package com.tanja.web_customer_tracker.model.customer;
 
 public enum GenderEnum {
 	MALE("m"),
