+++ src/main/java/com/tanja/web_customer_tracker/model/project/Project.java
@@ -1,4 +1,4 @@
-package com.tanja.web_customer_tracker.model;
+package com.tanja.web_customer_tracker.model.project;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -14,6 +14,8 @@ import javax.persistence.JoinColumn;
 import javax.persistence.JoinTable;
 import javax.persistence.ManyToMany;
 import javax.persistence.Table;
+
+import com.tanja.web_customer_tracker.model.bug.Bug;
 
 @Entity
 @Table(name = "project")
