+++ src/main/java/com/tanja/web_customer_tracker/rest/customer/CustomerRestController.java
@@ -1,4 +1,4 @@
-package com.tanja.web_customer_tracker.rest;
+package com.tanja.web_customer_tracker.rest.customer;
 
 import java.util.List;
 
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.tanja.web_customer_tracker.model.Customer;
-import com.tanja.web_customer_tracker.service.CustomerService;
+import com.tanja.web_customer_tracker.model.customer.Customer;
+import com.tanja.web_customer_tracker.service.customer.CustomerService;
 
 @RestController
 @RequestMapping("/api")
