+++ src/main/java/com/tanja/web_customer_tracker/service/customer/CustomerServiceImpl.java
@@ -1,4 +1,4 @@
-package com.tanja.web_customer_tracker.service;
+package com.tanja.web_customer_tracker.service.customer;
 
 import java.util.List;
 
@@ -6,8 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.tanja.web_customer_tracker.DAO.CustomerDAO;
-import com.tanja.web_customer_tracker.model.Customer;
+import com.tanja.web_customer_tracker.DAO.customer.CustomerDAO;
+import com.tanja.web_customer_tracker.model.customer.Customer;
 
 @Service
 public class CustomerServiceImpl implements CustomerService {
